fix: extract useStake hook
try-catch check token
