+++ src/helpers/formatNumber.ts
@@ -1,7 +1,7 @@
 export function formatNumber(num: number | undefined, digits = 2) {
   if (!num) return num
   return num.toLocaleString('en-US', {
-    minimumFractionDigits: 2,
+    minimumFractionDigits: Math.min(2, digits),
     maximumFractionDigits: digits,
   })
 }
